refactor: extract speak() out of speech-store into speech/speak.ts
Stores don't speak. speech-store.ts now owns only the voice catalog
and persisted config; speak() moves to its own file, mirroring how
audio/play-audio.ts models playAudio.
