+++ src/shared/speech/speak.test.ts
@@ -1,13 +1,13 @@
 import { beforeEach, describe, expect, test, vi } from "vitest";
 
-describe("speech-store without Web Speech API", () => {
+describe("speak() without Web Speech API", () => {
   beforeEach(() => {
     vi.resetModules();
     vi.stubGlobal("speechSynthesis", undefined);
   });
 
   test("speak() resolves silently instead of throwing", async () => {
-    const { speak } = await import("./speech-store");
+    const { speak } = await import("./speak");
     await expect(speak("hello")).resolves.toBeUndefined();
   });
 });
